Add files via upload
Description- According to EuropeanBusinessReview, telecommunication providers lose close to $65 million a month from customer churn. Isn't that expensive? With many emerging telecom giants, the competition in the telecom sector is increasing and the chances of customers discontinuing a service are high. This is often referred to as Customer Churn in Telecom Telecommunication providers that focus on quality service, lower-cost subscription plans, availability of content and features whilst creating positive customer service experiences have high chances of customer retention. The good news is that all these factors can be measured with different layers of data about billing history, subscription plans, cost of content, network/bandwidth utilization, and more to get a 360-degree view of the customer. This 360-degree view of customer data can be leveraged for predictive analytics to identify patterns and various trends that influence customer satisfaction and help reduce churn in telecom.


Expected Outcomes:

Accurate prediction of customer churn likelihood.
Insights into factors contributing to churn in the telecom industry.
Informed decision-making for proactive customer retention.
Tools and Technologies:

Python programming language.

Libraries such as pandas, scikit-learn for data manipulation and machine learning.
Jupyter Notebooks for interactive data analysis.
GitHub for version control and collaboration.


This project aims to empower telecom industry stakeholders with a predictive tool to anticipate and mitigate customer churn, ultimately contributing to improved customer satisfaction and business sustainability. The codebase, documentation, and results are shared on GitHub to facilitate collaboration and knowledge dissemination in the data science community.